Complete and update mailroom part 4 and mailroom test
